refactor: split the flat crate into app and ui crates
