+++ crates/app/src/assets.rs
