+++ crates/app/src/lib.rs
@@ -1,0 +1,7 @@
+//! Cydonia — a desktop client for ACP agents.
+
+pub mod acp;
+pub mod agents;
+pub mod assets;
+pub mod model;
+pub mod view;
