+++ crates/app/src/model/mod.rs
@@ -1,0 +1,9 @@
+//! What the app is: the projects that are open, the sessions running in them,
+//! the boards beside them, and the files all of that is restored from.
+
+pub mod board;
+pub mod project;
+pub mod session;
+pub mod settings;
+pub mod state;
+pub mod workspace;
