+++ crates/app/src/model/session.rs
@@ -8,8 +8,8 @@
 
 use crate::{
     acp::{self, Event, Reply, Session},
-    app::Cydonia,
-    settings, transcript,
+    model::{settings, workspace::Workspace},
+    view::transcript,
 };
 use anyhow::anyhow;
 use bezel::motion::Painter;
@@ -86,7 +86,7 @@ pub struct ChatSession {
     pub lost: bool,
     pub queue: VecDeque<String>,
     /// A prompt to send the moment the session is up — the card that opened
-    /// it. Taken by [`crate::app::Cydonia::session_connected`], never resent.
+    /// it. Taken by [`crate::view::root::Cydonia::session_connected`], never resent.
     pub seed: Option<String>,
     pub transcript: transcript::State,
     _pump: Task<()>,
@@ -98,7 +98,7 @@ impl ChatSession {
         entry: settings::Agent,
         cwd: PathBuf,
         seed: Option<String>,
-        cx: &mut Context<Cydonia>,
+        cx: &mut Context<Workspace>,
     ) -> Self {
         let spawn_entry = entry.clone();
         let conn = acp::runtime()
@@ -111,8 +111,8 @@ impl ChatSession {
             let (session, mut events) = match opened {
                 Ok(pair) => pair,
                 Err(e) => {
-                    let _ = this.update(cx, |app, cx| {
-                        app.with_session(id, cx, |chat| {
+                    let _ = this.update(cx, |workspace, cx| {
+                        workspace.with_session(id, cx, |chat| {
                             chat.lost = true;
                             chat.notice(true, &format!("connection failed: {e:#}"));
                         });
@@ -122,9 +122,9 @@ impl ChatSession {
             };
 
             if this
-                .update(cx, |app, cx| {
-                    app.with_session(id, cx, |chat| chat.session = Some(session));
-                    app.session_connected(id, cx);
+                .update(cx, |workspace, cx| {
+                    workspace.with_session(id, cx, |chat| chat.session = Some(session));
+                    workspace.session_connected(id, cx);
                 })
                 .is_err()
             {
@@ -136,13 +136,13 @@ impl ChatSession {
                 while let Ok(event) = events.try_recv() {
                     batch.push(event);
                 }
-                let streaming = this.update(cx, |app, cx| {
-                    app.with_session(id, cx, |chat| {
+                let streaming = this.update(cx, |workspace, cx| {
+                    workspace.with_session(id, cx, |chat| {
                         for event in batch {
                             chat.apply(event);
                         }
                     });
-                    app.session(id).is_some_and(|chat| chat.streaming)
+                    workspace.session(id).is_some_and(|chat| chat.streaming)
                 });
                 match streaming {
                     Ok(true) => cx.background_executor().timer(STREAM_FRAME).await,
