+++ crates/app/src/view/composer.rs
@@ -260,18 +260,11 @@ impl Composer {
                 .relative()
                 .flex_none()
                 .child(
-                    div()
-                        .id("composer-agent")
+                    ui::ghost(theme, "composer-agent")
                         .px(px(8.))
                         .py(px(3.))
-                        .rounded(px(Theme::control_radius()))
                         .text_size(px(12.))
                         .text_color(theme.text_muted)
-                        .cursor_pointer()
-                        .hover(|el| el.bg(theme.glass_hover()))
-                        .flex()
-                        .flex_row()
-                        .items_center()
                         .gap(px(4.))
                         .children(agent.icon.map(|path| {
                             svg()
