+++ crates/ui/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "cydonia-ui"
+description = "Widgets cydonia grew that bezel has not got yet"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+bezel = "0.1.3"
+gpui = { version = "0.3.6", package = "bezel-gpui" }
