--- src/lib.rs
@@ -1,14 +1,0 @@
-//! Cydonia — a desktop client for ACP agents.
-
-pub mod acp;
-pub mod agents;
-pub mod app;
-pub mod assets;
-pub mod board;
-pub mod composer;
-pub mod project;
-pub mod session;
-pub mod settings;
-pub mod settings_window;
-pub mod state;
-pub mod transcript;
